requirements.txt
fastapi
python-multipart
SQLAlchemy
PyMySQL
python-dotenv
uvicorn
Jinja2
pydantic-settings
redis